% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_msf_pbp.R
\name{add_score_data}
\alias{add_score_data}
\title{Add score data to custom pbp data}
\usage{
add_score_data(custom)
}
\arguments{
\item{custom}{a data.frame of custom pbp data}
}
\value{
the same data.frame as the input with multiple added columns
}
\description{
Add score data to custom pbp data
}
