% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp_data.R
\name{add_segment_data}
\alias{add_segment_data}
\title{Add segment information to custom play-by-play. This creates a segement id for every row
A segment is a contiguous period of time with the same 5 players on the court.}
\usage{
add_segment_data(pof, tes)
}
\description{
Add segment information to custom play-by-play. This creates a segement id for every row
A segment is a contiguous period of time with the same 5 players on the court.
}
