% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_per.R
\name{adjust_uper}
\alias{adjust_uper}
\title{Convert the uPER rating for a player to PER}
\usage{
adjust_uper(uper, lg.pace, tm.pace, lgu.per)
}
\arguments{
\item{uper}{the unadjusted per for the player}

\item{lg.pace}{league pace (possesions per game)}

\item{tm.pace}{team pace (possesions per game)}

\item{lgu.per}{league PER average}
}
\description{
Convert the uPER rating for a player to PER
}
