% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_audit_functions.R
\name{audit_pof_vec}
\alias{audit_pof_vec}
\title{Audit the gs_this_pof_vec column for pbp data printing out informative messages to the console}
\usage{
audit_pof_vec(pbp, team, pt, games)
}
\arguments{
\item{pbp}{a tibble of play-by-play data}

\item{team}{desired team's name, nickname, abbreviation or id}

\item{pt}{a data.frame of playing time from the game in question, as returned by the function \code{compute_game_playing_time()}}
}
\description{
Audit the gs_this_pof_vec column for pbp data printing out informative messages to the console
}
\seealso{
Other pbp_audit: 
\code{\link{compute_game_playing_time}()}
}
\concept{pbp_audit}
