% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup_helpers.R
\name{break_lineup_id}
\alias{break_lineup_id}
\title{break a dash separated id string into its component parts and return as a vector}
\usage{
break_lineup_id(lineup_id)
}
\arguments{
\item{lineup_id}{a string of dash separated player ids}
}
\value{
a vector of idividual ids
}
\description{
break a dash separated id string into its component parts and return as a vector
}
