% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_lineup_performance.R
\name{compare_lineups}
\alias{compare_lineups}
\title{Given a set of pbp data and player id inclusions and exclusions for 2 different lineup deinfitions, return a lineup summary comparing the two provided lineups}
\usage{
compare_lineups(pbp, lineup1, lineup2, round = 4)
}
\arguments{
\item{pbp}{a pbp dataset}

\item{lineup1}{a length two list with names "includes" and "excludes" defining the first lineup}

\item{lineup2}{a length two list with names "includes" and "excludes" defining the second lineup}
}
\value{
a 2 row data.frame of summarized pbp data, one row for each id
}
\description{
Any overlapping rows are removed and a summary of overlapping minutes is returned.
}
\seealso{
Other summarize_lineup: 
\code{\link{summarize_lineup_performance}()}
}
\concept{summarize_lineup}
