% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{compute_average_plus_minus}
\alias{compute_average_plus_minus}
\title{Compute the average plus minus for the reference team for a given chunk of pbp data}
\usage{
compute_average_plus_minus(dat, units = "minute")
}
\arguments{
\item{dat}{pbp data}

\item{units}{must be one of 'second', 'minute' (the default), 'game'}
}
\value{
the average plus minus for the reference team for the provided data
}
\description{
Compute the average plus minus for the reference team for a given chunk of pbp data
}
\seealso{
Other summarize_pbp_data: 
\code{\link{compute_defensive_free_throw_rate}()},
\code{\link{compute_defensive_rebound_rate}()},
\code{\link{compute_defensive_turnover_rate}()},
\code{\link{compute_effective_defensive_fgp}()},
\code{\link{compute_effective_fgp}()},
\code{\link{compute_free_throw_rate}()},
\code{\link{compute_plus_minus}()},
\code{\link{compute_rebound_rate}()},
\code{\link{compute_turnover_rate}()},
\code{\link{estimate_team_possessions_basic}()},
\code{\link{estimate_team_possessions_custom}()}
}
\concept{summarize_pbp_data}
