% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{compute_defensive_turnover_rate}
\alias{compute_defensive_turnover_rate}
\title{Compute the defensive turnover rate for a given period of pbp time}
\usage{
compute_defensive_turnover_rate(dat)
}
\arguments{
\item{dat}{pbp data}
}
\value{
defensive turnover rate for the given chunk of pbp data
}
\description{
Compute the defensive turnover rate for a given period of pbp time
}
\seealso{
Other summarize_pbp_data: 
\code{\link{compute_average_plus_minus}()},
\code{\link{compute_defensive_free_throw_rate}()},
\code{\link{compute_defensive_rebound_rate}()},
\code{\link{compute_effective_defensive_fgp}()},
\code{\link{compute_effective_fgp}()},
\code{\link{compute_free_throw_rate}()},
\code{\link{compute_plus_minus}()},
\code{\link{compute_rebound_rate}()},
\code{\link{compute_turnover_rate}()},
\code{\link{estimate_team_possessions_basic}()},
\code{\link{estimate_team_possessions_custom}()}
}
\concept{summarize_pbp_data}
