% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{compute_effective_defensive_fgp}
\alias{compute_effective_defensive_fgp}
\title{Compute defensive effective field goal percentage}
\usage{
compute_effective_defensive_fgp(dat)
}
\arguments{
\item{dat}{a tibble with play-by-play data}
}
\value{
the defensive effective field goal percentage for a given chunk of pbp data for the reference team
}
\description{
This function returns both the offensive and defensive effective field goal
percentage for the reference team in a given chunk of pbp data
}
\seealso{
Other summarize_pbp_data: \code{\link{compute_average_plus_minus}},
  \code{\link{compute_defensive_free_throw_rate}},
  \code{\link{compute_defensive_rebound_rate}},
  \code{\link{compute_defensive_turnover_rate}},
  \code{\link{compute_effective_fgp}},
  \code{\link{compute_free_throw_rate}},
  \code{\link{compute_plus_minus}},
  \code{\link{compute_rebound_rate}},
  \code{\link{compute_turnover_rate}},
  \code{\link{estimate_team_possessions_basic}},
  \code{\link{estimate_team_possessions_custom}}
}
\concept{summarize_pbp_data}
