% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{compute_effective_field_goal_percentage}
\alias{compute_effective_field_goal_percentage}
\title{Compute effective field goal percentage}
\usage{
compute_effective_field_goal_percentage(dat, team = "this")
}
\arguments{
\item{dat}{a tibble with play-by-play data}
}
\value{
a dbl of a team's effective field goal percentage for the period in question
}
\description{
Compute effective field goal percentage
}
