% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{compute_eight_factors}
\alias{compute_eight_factors}
\title{This function wraps the eight factors functions and returns a tibble with a single
row of all eight factors}
\usage{
compute_eight_factors(dat, ...)
}
\arguments{
\item{dat}{a chunk of pbp data}
}
\value{
a one_row tibble with 8 factors data for the given input
chunk of pbp data
}
\description{
This function wraps the eight factors functions and returns a tibble with a single
row of all eight factors
}
\seealso{
Other summarize_pbp_data: \code{\link{compute_effective_field_goal_percentage}},
  \code{\link{compute_free_throw_rate}},
  \code{\link{compute_rebound_rate}},
  \code{\link{compute_turnover_rate}},
  \code{\link{estimate_team_possessions_basic}},
  \code{\link{estimate_team_possessions_custom}}
}
