% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{compute_free_throw_rate}
\alias{compute_free_throw_rate}
\title{Compute the free throw rate for a chunk of pbp data}
\usage{
compute_free_throw_rate(dat, use_attempted = TRUE)
}
\arguments{
\item{dat}{a tibble of pbp data}
}
\value{
the free throw rate for the reference team for the given chunk of pbp data
}
\description{
Compute the free throw rate for a chunk of pbp data
}
\seealso{
Other summarize_pbp_data: 
\code{\link{compute_average_plus_minus}()},
\code{\link{compute_defensive_free_throw_rate}()},
\code{\link{compute_defensive_rebound_rate}()},
\code{\link{compute_defensive_turnover_rate}()},
\code{\link{compute_effective_defensive_fgp}()},
\code{\link{compute_effective_fgp}()},
\code{\link{compute_plus_minus}()},
\code{\link{compute_rebound_rate}()},
\code{\link{compute_turnover_rate}()},
\code{\link{estimate_team_possessions_basic}()},
\code{\link{estimate_team_possessions_custom}()}
}
\concept{summarize_pbp_data}
