% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_audit_functions.R
\name{compute_game_playing_time}
\alias{compute_game_playing_time}
\title{Compute playing time for all players for a given game and team. This is used primarily to trouble shoot
players-on-the-floor issues caused by missing or extra substitutions}
\usage{
compute_game_playing_time(pbp, team)
}
\arguments{
\item{pbp}{play by play data for exactly one full game}

\item{team}{the city, nickname, abbreviation or id of the desired team}
}
\value{
a tibble indicating player id, last and first name, jersey number and minutes played
}
\description{
Compute playing time for all players for a given game and team. This is used primarily to trouble shoot
players-on-the-floor issues caused by missing or extra substitutions
}
\seealso{
Other pbp_audit: \code{\link{audit_pof_vec}}
}
