% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_per.R
\name{compute_league_drbd}
\alias{compute_league_drbd}
\title{Compute league wide defensive rebounding percentage}
\usage{
compute_league_drbd(lg.trb, lg.orb)
}
\arguments{
\item{lg.trb}{total league rebounds}

\item{lg.orb}{total league offensive rebounds}
}
\description{
Compute league wide defensive rebounding percentage
}
