% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_per.R
\name{compute_league_vop}
\alias{compute_league_vop}
\title{Compute the league value of play (for the PER rating - see \code{compute_per()})}
\usage{
compute_league_vop(lg.pts, lg.fga, lg.orb, lg.to, lg.fta)
}
\arguments{
\item{lg.pts}{total leaguge points scored}

\item{lg.fga}{number of league field goals attempted}

\item{lg.orb}{number of league offensive rebounds}

\item{lg.to}{number of league turnovers}

\item{lg.fta}{number of league free throw attempts}
}
\description{
Compute the league value of play (for the PER rating - see \code{compute_per()})
}
