% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{compute_rebound_rate}
\alias{compute_rebound_rate}
\title{Compute the offensive and defensive rebound rate for the reference team for a given
chunk of pbp data}
\usage{
compute_rebound_rate(dat)
}
\arguments{
\item{dat}{a tibble with pbp data}
}
\value{
a list with orp (offensive rebounding percentage) and drp (defensive rebounding percentage)
}
\description{
Compute the offensive and defensive rebound rate for the reference team for a given
chunk of pbp data
}
\seealso{
Other summarize_pbp_data: \code{\link{compute_effective_field_goal_percentage}},
  \code{\link{compute_eight_factors}},
  \code{\link{compute_free_throw_rate}},
  \code{\link{compute_turnover_rate}},
  \code{\link{estimate_team_possessions_basic}},
  \code{\link{estimate_team_possessions_custom}}
}
