% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_summary_functions.R
\name{compute_team_eight_factors}
\alias{compute_team_eight_factors}
\title{Compute the 8 factors for all games for a given team}
\usage{
compute_team_eight_factors(team)
}
\arguments{
\item{team}{a team name, city, abrreviation or id}
}
\value{
a tibble with team_id, game_id, game_number and each of the 8 factors
}
\description{
Compute the 8 factors for all games for a given team
}
