% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_team_stats.R
\name{compute_team_stats}
\alias{compute_team_stats}
\title{Compute basic team level stats from play-by-play data}
\usage{
compute_team_stats(pbp_data)
}
\arguments{
\item{pbp_data}{a chunk of play-by-play data}
}
\value{
a tibble summarizing basic team level stats typically found in a box score
}
\description{
Compute the number of points scored by a team and its opponent in a
given chunk of pbp data
}
