% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_msf_pbp.R
\name{compute_tes}
\alias{compute_tes}
\title{compute the total time elapsed in the game given the quarter and time elapsed}
\usage{
compute_tes(x, y)
}
\description{
compute the total time elapsed in the game given the quarter and time elapsed
}
