% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{compute_turnover_rate}
\alias{compute_turnover_rate}
\title{Compute the turnover rate for a given period of pbp time}
\usage{
compute_turnover_rate(dat, team = "this")
}
\arguments{
\item{dat}{pbp data}
}
\value{
a dbl of team's turnover rate
}
\description{
Compute the turnover rate for a given period of pbp time
}
