% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_per.R
\name{compute_uper}
\alias{compute_uper}
\title{Compute the PER (player efficiency ranking) for a single player
see https://en.wikipedia.org/wiki/Player_efficiency_rating}
\usage{
compute_uper(thp, pf, min, fg, ft, fta, ast, lg.ft, lg.pf, tm.ast, tm.fg, orb,
  trb, to, vop, drbp, factor)
}
\arguments{
\item{thp}{number of three pointers made}

\item{pf}{number of personal fouls}

\item{min}{number of minutes played}

\item{fg}{number of field goals made}

\item{ft}{number of free throws made}

\item{fta}{number of free throws attempted}

\item{ast}{number of assists}

\item{lg.ft}{number of league free throws}

\item{lg.pf}{number of league personal fouls}

\item{tm.ast}{number of team assists}

\item{tm.fg}{number of team field goals}

\item{orb}{number of offensive rebounds}

\item{trb}{number of total rebounds}

\item{to}{number of turnovers}

\item{vop}{value of play (see \code{compute_leauge_vop()})}

\item{drbp}{league defensize rebound percentage (see \code{compute_league_drbp})}

\item{factor}{league based adjustment factor based on assists, filed goals and free throws}
}
\description{
Compute the PER (player efficiency ranking) for a single player
see https://en.wikipedia.org/wiki/Player_efficiency_rating
}
