% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_msf_pbp.R
\name{customize_msf_pbp}
\alias{customize_msf_pbp}
\title{Transform raw pbp data so that it is specific to one of the teams}
\usage{
customize_msf_pbp(plays, team_id, opponent_id, loc)
}
\arguments{
\item{plays}{a raw data.frame from msf play-by-play data}

\item{team_id}{id of the reference team}

\item{opponent_id}{id of the opponent team}

\item{loc}{either 'home' or 'away' indicating the location from the
perspective of the reference team}
}
\value{
a data.frame with customized pbp data
}
\description{
Used internally by process_msf_pbp()
}
