% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp_data.R
\name{customize_pbp}
\alias{customize_pbp}
\title{a helper function for process_raw_pbp that converts a raw play-by-play data.frame to one customized to
the home or away team}
\usage{
customize_pbp(pbp, team_id, opponent_id, loc)
}
\arguments{
\item{pbp}{raw pbp data}

\item{team_id}{msf id of the team you wish to customize the report for}

\item{opponent_id}{msf id of the opponent}

\item{loc}{location of game with reference to custom team - either 'home' or 'away'}
}
\value{
a play-by-play customized to one team with references to 'home' and 'away' changed to 'team' and 'opponent'
accordingly
}
\description{
a helper function for process_raw_pbp that converts a raw play-by-play data.frame to one customized to
the home or away team
}
