% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{estimate_possessions_advanced}
\alias{estimate_possessions_advanced}
\title{estimate the number of possessions for a given chunk of play-by-play data using the advanced method
see https://www.nbastuffer.com/analytics101/possession/}
\usage{
estimate_possessions_advanced(fga, fta, or, fg, to, dr, opp_fga, opp_fta,
  opp_or, opp_fg, opp_to, opp_dr)
}
\arguments{
\item{fga}{number of field goal attempts in the period}

\item{fta}{free throw attempts}

\item{or}{offensive rebounds}

\item{fg}{field goals}

\item{to}{turnovers}

\item{dr}{defensize rebounds}

\item{opp_fga}{opponent field goal attempts}

\item{opp_fta}{opponent free throw attempts}

\item{opp_or}{opponent offensize rebounds}

\item{opp_fg}{opponent field goals}

\item{opp_to}{opponent turnovers}

\item{opp_dr}{opponent defensive rebounds}
}
\value{
the estimated number of possessions for the period
}
\description{
estimate the number of possessions for a given chunk of play-by-play data using the advanced method
see https://www.nbastuffer.com/analytics101/possession/
}
