% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{estimate_possessions_basic}
\alias{estimate_possessions_basic}
\title{Estimate the number of possessions for a given chunk of play-by-play data using the basic method
see https://www.nbastuffer.com/analytics101/possession/}
\usage{
estimate_possessions_basic(dat)
}
\arguments{
\item{fga}{number of field goal attempts in the period}

\item{to}{turnovers}

\item{fta}{free throw attempts}

\item{or}{offensive rebounds}
}
\value{
the estimated number of possessions for the period
}
\description{
Estimate the number of possessions for a given chunk of play-by-play data using the basic method
see https://www.nbastuffer.com/analytics101/possession/
}
