% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup_helpers.R
\name{filter_lineup}
\alias{filter_lineup}
\title{a function for searching lineups that do or do not include certain players}
\usage{
filter_lineup(lineup, inlcudes, excludes)
}
\description{
a function for searching lineups that do or do not include certain players
}
