% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_lineup.R
\name{fix_lineup}
\alias{fix_lineup}
\title{This funtion is used to fix erroneous lineups on load}
\usage{
fix_lineup(lu, game_id, team_id)
}
\description{
This funtion is used to fix erroneous lineups on load
}
