% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_audit_functions.R
\name{generate_substitution_row}
\alias{generate_substitution_row}
\title{generate a row of substitution data for a pbp data.frame. This can be dput into fix pbp to add a missing row}
\usage{
generate_substitution_row(
  player_in_id,
  player_out_id,
  player_data,
  quarter,
  elapsed_time_in_quarter
)
}
\arguments{
\item{player_in_id}{the id of the player entering the game}

\item{player_out_id}{the id of the player leaving the game}

\item{player_data}{a data.frame of player data, as returned by \code{get_player_data()$api_json$players}}

\item{quarter}{the quarter in which the substitution should occur}

\item{elapsed_time_in_quarter}{number of seconds since the start of the quarter when the sub should be inserted}
}
\value{
a one row data.frame with substitution data. This can be inserted into a pbp data.frame
}
\description{
generate a row of substitution data for a pbp data.frame. This can be dput into fix pbp to add a missing row
}
\seealso{
Other audit_functions: 
\code{\link{generate_substitution_row_from_desc}()},
\code{\link{get_player_ids_from_desc}()},
\code{\link{read_boxscore_screen_shot}()},
\code{\link{remove_pbp}()}
}
\concept{audit_functions}
