% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pbp.R
\name{get_all_completed_game_ids}
\alias{get_all_completed_game_ids}
\title{Load a vector of all completed game_ids}
\usage{
get_all_completed_game_ids()
}
\value{
a vector of all completed game_ids
}
\description{
Load a vector of all completed game_ids
}
