% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{get_lineup}
\alias{get_lineup}
\title{Get a list of starters and bench players for a given game_id}
\usage{
get_lineup(game_id, team)
}
\arguments{
\item{game_id}{the MSF game id for the desired game}

\item{team}{strng to indicate team. Is passed to interpret team so can be any of name, city, id or abbreviation}
}
\description{
Get a list of starters and bench players for a given game_id
}
