% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup_helpers.R
\name{get_lineup_initials}
\alias{get_lineup_initials}
\title{Convert a lineup to a string of player initials + player jersey number}
\usage{
get_lineup_initials(lineup, pd)
}
\arguments{
\item{lineup}{a lineup vector}

\item{pd}{player data}
}
\value{
a concateneated string of initial-number combinations
}
\description{
Convert a lineup to a string of player initials + player jersey number
}
\seealso{
Other lineup_helpers: 
\code{\link{compare_lineups}()},
\code{\link{get_lineup_last_names}()},
\code{\link{subset_pbp_by_player_ids}()},
\code{\link{tag_pbp_by_player_ids}()}
}
\concept{lineup_helpers}
