% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{get_player_game_data}
\alias{get_player_game_data}
\title{Load the raw game box score data and return a tibble of data for all players
for a given team in that game}
\usage{
get_player_game_data(game_id, team)
}
\arguments{
\item{game_id}{msf game id}

\item{team}{the msf team id, name, city or abbr for the desired team's data}
}
\value{
raw box score data for all players on the requested team for the requested game
}
\description{
Load the raw game box score data and return a tibble of data for all players
for a given team in that game
}
