% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_audit_functions.R
\name{get_player_ids_from_desc}
\alias{get_player_ids_from_desc}
\title{Given a sentence of data from ESPN play by play about a substituon, return the player ids of the entering and exiting player. Note that this function only works when the players have a single first and last name (e.g. Julius Randle, but not J.R. Barrett)}
\usage{
get_player_ids_from_desc(sen, player_data)
}
\arguments{
\item{sen}{the text indicating the substition (e.g."Mitchell Robinson enters the game for Julius Randle")}

\item{player_data}{a data.frame of player data, as returned by \code{get_player_data()$api_json$players}}
}
\value{
a length two list with names "pi_id" and "po_id" giving the msf player ids of the two players involved in the substitution
}
\description{
Given a sentence of data from ESPN play by play about a substituon, return the player ids of the entering and exiting player. Note that this function only works when the players have a single first and last name (e.g. Julius Randle, but not J.R. Barrett)
}
\seealso{
Other audit_functions: 
\code{\link{generate_substitution_row_from_desc}()},
\code{\link{generate_substitution_row}()},
\code{\link{read_boxscore_screen_shot}()}
}
\concept{audit_functions}
