% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{get_raw_box_score}
\alias{get_raw_box_score}
\title{Get the box score for a given game id}
\usage{
get_raw_box_score(game_id, use_archive = TRUE)
}
\arguments{
\item{game_id}{the msf game id for the requested game}

\item{use_archive}{when TRUE (the default) will archive the raw download to tidynbadata$ARCHIVE_DIR}
}
\description{
Get the box score for a given game id
}
