% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{get_raw_msf_box_score}
\alias{get_raw_msf_box_score}
\title{Retrieve a raw box score pull from mysportsfeeds based on a game_id}
\usage{
get_raw_msf_box_score(game_id)
}
\arguments{
\item{game_id}{}
}
\value{
a raw boxscore object from mysportsfeeds, which is also archived
}
\description{
Retrieve a raw box score pull from mysportsfeeds based on a game_id
}
