% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{get_raw_pbp}
\alias{get_raw_pbp}
\title{Get a raw copy of a game's play-by-play data from the local archive when possible}
\usage{
get_raw_pbp(game_id)
}
\arguments{
\item{game_id}{the msf game id for the game}
}
\value{
a raw play-by-play object from mysportsfeeds, which is also archived
}
\description{
Get a raw copy of a game's play-by-play data from the local archive when possible
}
\seealso{
Other data_acquisition: 
\code{\link{get_lineup}()},
\code{\link{get_player_data}()},
\code{\link{get_raw_msf_box_score}()},
\code{\link{get_team_schedule}()}
}
\concept{data_acquisition}
