% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{get_team_game_data}
\alias{get_team_game_data}
\title{Load the raw team box score data and return a tibble of data for team data}
\usage{
get_team_game_data(game_id)
}
\arguments{
\item{game_id}{msf game id}

\item{team}{the msf team id, name, city or abbr for the desired team's data}
}
\value{
raw box score data for the requested team for the requested game
}
\description{
Load the raw team box score data and return a tibble of data for team data
}
