% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tidynba_names.R
\name{load_pbp_name_map}
\alias{load_pbp_name_map}
\title{Generates a name_map that converts the names of a raw play-by-play data.frame
to the names expected by tidynbadata}
\usage{
load_pbp_name_map()
}
\description{
Generates a name_map that converts the names of a raw play-by-play data.frame
to the names expected by tidynbadata
}
