% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup_helpers.R
\name{make_lineup_id}
\alias{make_lineup_id}
\title{convert a vector of player ids into a single dash separated string, sorted
low to high}
\usage{
make_lineup_id(lineup)
}
\arguments{
\item{lineup}{a vector of player ids}
}
\value{
a single string of each provided id separated by dashed or NA if
length of provided vector is not 5
}
\description{
convert a vector of player ids into a single dash separated string, sorted
low to high
}
