% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_game_score}
\alias{plot_game_score}
\title{Plot the score differential of a game based on play-by-play data}
\usage{
plot_game_score(pbp)
}
\arguments{
\item{pbp}{a data.frame of play-by-play data from a single game}
}
\value{
a ggplot object
}
\description{
Plot the score differential of a game based on play-by-play data
}
