% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_lineup_comparision}
\alias{plot_lineup_comparision}
\title{Plot the output of \code{compare_lineups()} to display the statistical differences between the lineups}
\usage{
plot_lineup_comparision(lineup_data)
}
\arguments{
\item{lineup_data}{data.frame as returned by \code{compare_lineups()}}
}
\value{
a list with game data and multiple plots of lineup data
}
\description{
Plot the output of \code{compare_lineups()} to display the statistical differences between the lineups
}
\concept{plotting_functions}
