% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_msf_pbp.R
\name{process_lineups}
\alias{process_lineups}
\title{a helper function for \code{process_msf_pbp()} that adds list columns indicating which players
are on the floor for any moment of the game}
\usage{
process_lineups(team, subs, game_id)
}
\arguments{
\item{team}{team id, name, city or abbreviation}

\item{subs}{data.frame of substitution information}
}
\value{
play-by-play data with players_on_the_floor columns attached
}
\description{
a helper function for \code{process_msf_pbp()} that adds list columns indicating which players
are on the floor for any moment of the game
}
