% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_msf_pbp.R
\name{process_msf_pbp}
\alias{process_msf_pbp}
\title{High level function for converting raw msf pbp to a tidynbadata pbp tibble}
\usage{
process_msf_pbp(raw_msf_pbp)
}
\arguments{
\item{game_id}{}
}
\value{
a length 2 list of pbp tibbles, one for each team involved.
This list is also archived.
}
\description{
High level function for converting raw msf pbp to a tidynbadata pbp tibble
}
