% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp_data.R
\name{process_raw_pbp}
\alias{process_raw_pbp}
\title{Functions for processing and customizing play by play data}
\usage{
process_raw_pbp(game_id)
}
\description{
This function replaces add_lineups_to_pbp. It uses the raw results of the
msf api call, extracts the play-by-play archive data.frame and:
\enumerate{
\item adds players on the floor vectors for each team
\item adds columns - OriginalRow this just numbers each row in the original data.frame for later rejoins
- total_elapsed_seconds - this indicates the number of seconds of play time elapsed since
the beginning of the game
\item personalizes the data.frame for each of the two teams involved, this means:
a) rename the columns from "home"-"away" to "team"-"opponent"
b) does the same with the players on the floor list columns
\item archives both versions of the pbp in the tidynbadata$ARCHIVE_DIR/proc_pbp_archive
\item Add other columns (i.e. current team score, current oppononet score, game_status (winning, losing, tied))
}
}
