% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_msf_pbp.R
\name{process_sub}
\alias{process_sub}
\title{a helper function for inserting and removing players from a lineup vector}
\usage{
process_sub(current, player_in, player_out)
}
\arguments{
\item{current}{the vector of player ids currently in the game}

\item{player_in}{the id of the player coming into the game}

\item{player_out}{the id of the player leaving the game}
}
\value{
a low to high sorted vector of the player ids on the court after the
substitution is complete
}
\description{
a helper function for inserting and removing players from a lineup vector
}
