% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_audit_functions.R
\name{read_boxscore_screen_shot}
\alias{read_boxscore_screen_shot}
\title{convert a screen shot of ESPN playing time data to a data.frame}
\usage{
read_boxscore_screen_shot(infile = "most_recent_screen_shot")
}
\arguments{
\item{infile}{the path to the screen shot, by default will use the most recent screen shot on the Desktop}
}
\value{
data.frame of playing time info
}
\description{
convert a screen shot of ESPN playing time data to a data.frame
}
\seealso{
Other audit_functions: 
\code{\link{generate_substitution_row_from_desc}()},
\code{\link{generate_substitution_row}()},
\code{\link{get_player_ids_from_desc}()}
}
\concept{audit_functions}
