% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_audit_functions.R
\name{remove_pbp}
\alias{remove_pbp}
\title{Delete an archived processed play-by-play data set. Typically this is done when we know that it contains errors and want to force a reprocessing}
\usage{
remove_pbp(game_id)
}
\value{
a Boolean indicting if the remove operation was successful
}
\description{
Delete an archived processed play-by-play data set. Typically this is done when we know that it contains errors and want to force a reprocessing
}
\seealso{
Other audit_functions: 
\code{\link{generate_substitution_row_from_desc}()},
\code{\link{generate_substitution_row}()},
\code{\link{get_player_ids_from_desc}()},
\code{\link{read_boxscore_screen_shot}()}
}
\concept{audit_functions}
