% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tidynba_names.R
\name{set_tidynba_names}
\alias{set_tidynba_names}
\title{A function that converts the names of a data.frame to those provided by a name_map.
The name_map is a named vector in which the names are the column names of the raw data
and the values are the column names expected by tidynbadata functions}
\usage{
set_tidynba_names(dat, name_map)
}
\arguments{
\item{dat}{a tibble (or data.frame)}

\item{name_map}{a named vector the names of which should match the names of \code{dat}}
}
\value{
dat, with new names as provided by the values of the name_map
}
\description{
A function that converts the names of a data.frame to those provided by a name_map.
The name_map is a named vector in which the names are the column names of the raw data
and the values are the column names expected by tidynbadata functions
}
