% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup_helpers.R
\name{subset_pbp_by_player_ids}
\alias{subset_pbp_by_player_ids}
\title{subset a pbp datset according to players that must be, or must not be, in the current lineup}
\usage{
subset_pbp_by_player_ids(pbp, includes, excludes)
}
\arguments{
\item{pbp}{a processed play-by-play dataset}

\item{includes}{a vector of player ids that must be in  \code{gs_this_pof_vec}}

\item{excludes}{a vector of player ids that myst not be in \code{gs_this_pof_vec}}
}
\value{
a pbp dataset subset according to includes and excludes
}
\description{
subset a pbp datset according to players that must be, or must not be, in the current lineup
}
\seealso{
Other lineup_helpers: 
\code{\link{compare_lineups}()},
\code{\link{get_lineup_initials}()},
\code{\link{get_lineup_last_names}()},
\code{\link{tag_pbp_by_player_ids}()}
}
\concept{lineup_helpers}
