% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_lineup_performance.R
\name{summarize_lineup_performance}
\alias{summarize_lineup_performance}
\title{Summarize the performance of each unique lineup in the provided pbp data}
\usage{
summarize_lineup_performance(dat, minimum_minutes)
}
\arguments{
\item{dat}{a tibble with play-by-play data}

\item{minimum_minutes}{lineups with fewer minutes played that this will be excluded}
}
\value{
a tibble with summary statistics for each unique lineup in the provided data
}
\description{
Summarize the performance of each unique lineup in the provided pbp data
}
