% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{summarize_raw_box_scores}
\alias{summarize_raw_box_scores}
\title{Transform raw box score data to a tidy format with new names}
\usage{
summarize_raw_box_scores(raw_box_score_data)
}
\arguments{
\item{box_score_data}{output from the function get_raw_box_score(game_id)}
}
\value{
a tibble with box score data for the game data provided
}
\description{
Transform raw box score data to a tidy format with new names
}
