% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup_helpers.R
\name{tag_pbp_by_player_ids}
\alias{tag_pbp_by_player_ids}
\title{tag a pbp datset according to players that must be, or must not be, in the current lineup}
\usage{
tag_pbp_by_player_ids(pbp, includes, excludes)
}
\arguments{
\item{pbp}{a processed play-by-play dataset}

\item{includes}{a vector of player ids that must be in  \code{gs_this_pof_vec}}

\item{excludes}{a vector of player ids that must not be in \code{gs_this_pof_vec}}
}
\value{
the same pbp dataset with a new Boolean column indicating which rows match the includes and excludes criteria
}
\description{
tag a pbp datset according to players that must be, or must not be, in the current lineup
}
\seealso{
Other lineup_helpers: 
\code{\link{get_lineup_initials}()},
\code{\link{get_lineup_last_names}()},
\code{\link{subset_pbp_by_player_ids}()}
}
\concept{lineup_helpers}
