% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{tidynbadata}
\alias{tidynbadata}
\alias{tidynbadata-package}
\alias{tidynbadata-package}
\title{Tools for computing advanced NBA statistics}
\description{
Tools for computing advanced NBA statistics
}
\section{Package Options}{


tidybnadata used the following options to configure behavior:

\itemize{
\item \code{tidynbadata.archive_path}: location of archive directory
\item \code{tidynbadata.current_season}: default season for msfsportsfeeds data
\item \code{tidynbadata.msf_version_id}: default version for msfsportsfeeds

}
}

\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
