% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{tidynbadata_team_info}
\alias{tidynbadata_team_info}
\title{Basic data on 30 NBA teams}
\format{a data.frame with 30 rows and 4 variables:
\describe{
\item{id}{two digit id (from mysportsfeeds) for identifying the team}
\item{city}{the city (or other place name) with which the team is associated}
\item{name}{The team's name (e.g. "Hawks", "Knicks", "Celtics")}
\item{abbr}{three capital letters that identify the team (e.g "ATL", "NYK", "BOS")}
}}
\source{
mysportsfeeds
}
\usage{
tidynbadata_team_info
}
\description{
a dataseat containing basic identification information for 30 NBA teams
}
\keyword{datasets}
