% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pbp.R
\name{update_pbp_summaries}
\alias{update_pbp_summaries}
\title{This function checks the archive to see which pbps have been processed
and the schedule to see which games have been played and then downloads,
processes and archives the play-by-play data, returning nothing.}
\usage{
update_pbp_summaries()
}
\description{
This function checks the archive to see which pbps have been processed
and the schedule to see which games have been played and then downloads,
processes and archives the play-by-play data, returning nothing.
}
