% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_pbp_stats.R
\name{verify_pbp_team_stats}
\alias{verify_pbp_team_stats}
\title{Compare the play-by-play computed team stats with the team stats provided by
mysportsfeeds as a quality check on the play-by-play data and its processing}
\usage{
verify_pbp_team_stats(game_id)
}
\arguments{
\item{game_id}{an msf game id}
}
\value{
NULL this function is used for its side effects
}
\description{
Compare the play-by-play computed team stats with the team stats provided by
mysportsfeeds as a quality check on the play-by-play data and its processing
}
