% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp_data.R
\name{add_score_data}
\alias{add_score_data}
\title{Add current score information to a customized play-by-play tibble}
\usage{
add_score_data(custom_pbp)
}
\arguments{
\item{custom_pbp}{}
}
\description{
Add current score information to a customized play-by-play tibble
}
