% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup_helpers.R
\name{compare_lineups}
\alias{compare_lineups}
\title{Given a set of pbp data and player id inclusions and exclusions for 2 different lineup deinfitions, return a lineup summary comparing the two provided lineups}
\usage{
compare_lineups(pbp, lineup1, lineup2, round = 4)
}
\arguments{
\item{pbp}{a pbp dataset}

\item{lineup1}{a length two list with names "includes" and "excludes" defining the first lineup}

\item{lineup2}{a length two list with names "includes" and "excludes" defining the second lineup}
}
\value{
a 2 row data.frame of summarized pbp data, one row for each id
}
\description{
Any overlapping rows are removed and a summary of overlapping minutes is returned.
}
\seealso{
Other lineup_helpers: 
\code{\link{get_lineup_initials}()},
\code{\link{get_lineup_last_names}()},
\code{\link{subset_pbp_by_player_ids}()},
\code{\link{tag_pbp_by_player_ids}()}
}
\concept{lineup_helpers}
