% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{compute_effective_field_goal_percentage}
\alias{compute_effective_field_goal_percentage}
\title{Compute effective field goal percentage}
\usage{
compute_effective_field_goal_percentage(dat)
}
\arguments{
\item{dat}{a tibble with play-by-play data}
}
\value{
a list with efgp (effective field goal percentage) and defgp (defensive effecgive field goal percentage)
}
\description{
This function returns both the offensive and defensive effective field goal
percentage for the reference team in a given chunk of pbp data
}
\seealso{
Other summarize_pbp_data: \code{\link{compute_eight_factors}},
  \code{\link{compute_free_throw_rate}},
  \code{\link{compute_rebound_rate}},
  \code{\link{compute_turnover_rate}},
  \code{\link{estimate_team_possessions_basic}},
  \code{\link{estimate_team_possessions_custom}}
}
\concept{summarize_pbp_data}
