% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{compute_free_throw_rate}
\alias{compute_free_throw_rate}
\title{Compute offensive and defensive rebound rate for a chunck of pbp data}
\usage{
compute_free_throw_rate(dat, use_attempted = TRUE)
}
\arguments{
\item{dat}{a tibble of pbp data}
}
\value{
a list with ftr (free throw rate) and dftr (defensive free throw rate)
}
\description{
Compute offensive and defensive rebound rate for a chunck of pbp data
}
\seealso{
Other summarize_pbp_data: \code{\link{compute_effective_field_goal_percentage}},
  \code{\link{compute_eight_factors}},
  \code{\link{compute_rebound_rate}},
  \code{\link{compute_turnover_rate}},
  \code{\link{estimate_team_possessions_basic}},
  \code{\link{estimate_team_possessions_custom}}
}
\concept{summarize_pbp_data}
