% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{compute_turnover_rate}
\alias{compute_turnover_rate}
\title{Compute the turnover rate for a given period of pbp time}
\usage{
compute_turnover_rate(dat)
}
\arguments{
\item{dat}{pbp data}
}
\value{
a list with reference team's turnover rate and defensive turnover rate
}
\description{
Compute the turnover rate for a given period of pbp time
}
\seealso{
Other summarize_pbp_data: \code{\link{compute_effective_field_goal_percentage}},
  \code{\link{compute_eight_factors}},
  \code{\link{compute_free_throw_rate}},
  \code{\link{compute_rebound_rate}},
  \code{\link{estimate_team_possessions_basic}},
  \code{\link{estimate_team_possessions_custom}}
}
