% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{estimate_team_possessions_basic}
\alias{estimate_team_possessions_basic}
\title{A basic function for estimating the number of possesions for a team
in a given chunk of play-by-play data}
\usage{
estimate_team_possessions_basic(dat)
}
\arguments{
\item{dat}{a chunk of pbp data}
}
\value{
a double estimating the number of plays in the provided pbp data
}
\description{
A basic function for estimating the number of possesions for a team
in a given chunk of play-by-play data
}
\seealso{
Other summarize_pbp_data: 
\code{\link{compute_average_plus_minus}()},
\code{\link{compute_defensive_free_throw_rate}()},
\code{\link{compute_defensive_rebound_rate}()},
\code{\link{compute_defensive_turnover_rate}()},
\code{\link{compute_effective_defensive_fgp}()},
\code{\link{compute_effective_fgp}()},
\code{\link{compute_free_throw_rate}()},
\code{\link{compute_plus_minus}()},
\code{\link{compute_rebound_rate}()},
\code{\link{compute_turnover_rate}()},
\code{\link{estimate_team_possessions_custom}()}
}
\concept{summarize_pbp_data}
