% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pbp_data.R
\name{estimate_team_possessions_custom}
\alias{estimate_team_possessions_custom}
\title{Estimate the number of possessions for a team for a given chink of pbp data}
\usage{
estimate_team_possessions_custom(dat)
}
\arguments{
\item{dat}{a chunk of pbp data to be analyzed}
}
\value{
estimated number of possesions in the pbp chunk
}
\description{
This function computes possesions as the sum of:
\enumerate{
\item opponent field goals - opponent offensive rebounds
\item opponent turnovers
\item opponent final free throws made (e.g the 2nd of 2 or 3rd of 3)
}
}
\seealso{
Other summarize_pbp_data: \code{\link{compute_effective_field_goal_percentage}},
  \code{\link{compute_eight_factors}},
  \code{\link{compute_free_throw_rate}},
  \code{\link{compute_rebound_rate}},
  \code{\link{compute_turnover_rate}},
  \code{\link{estimate_team_possessions_basic}}
}
\concept{summarize_pbp_data}
