% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{get_lineup}
\alias{get_lineup}
\title{Get a list of starters and bench players for a given game_id}
\usage{
get_lineup(game_id, team,
  season = getOption("tidynbadata.current_season"))
}
\arguments{
\item{game_id}{the MSF game id for the desired game}

\item{team}{strng to indicate team. Is passed to interpret team so can be any of name, city, id or abbreviation}

\item{season}{string name of desired season. Defaults to \code{getOption('tidynbadata.current_season')}}
}
\value{
tibble of lineup data for given game and team
}
\description{
Get a list of starters and bench players for a given game_id
}
\seealso{
Other data_acquisition: \code{\link{get_player_data}},
  \code{\link{get_raw_msf_box_score}},
  \code{\link{get_raw_pbp}},
  \code{\link{get_team_schedule}}
}
\concept{data_acquisition}
