% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{get_raw_msf_box_score}
\alias{get_raw_msf_box_score}
\title{Retrieve a raw box score pull from mysportsfeeds based on a game_id}
\usage{
get_raw_msf_box_score(game_id)
}
\arguments{
\item{game_id}{id of requested game}
}
\value{
a raw boxscore object from mysportsfeeds, which is also archived
}
\description{
Retrieve a raw box score pull from mysportsfeeds based on a game_id
}
\seealso{
Other data_acquisition: \code{\link{get_lineup}},
  \code{\link{get_player_data}}, \code{\link{get_raw_pbp}},
  \code{\link{get_team_schedule}}
}
\concept{data_acquisition}
