% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{get_raw_pbp}
\alias{get_raw_pbp}
\title{Get a raw copy of a game's play-by-play data from the local archive when possible}
\usage{
get_raw_pbp(game_id, use_archive = TRUE)
}
\arguments{
\item{game_id}{the msf game id for the game}

\item{use_archive}{when TRUE (the default) will first check locally for a copy and if not make a msf api call}
}
\description{
Get a raw copy of a game's play-by-play data from the local archive when possible
}
