% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{get_team_schedule}
\alias{get_team_schedule}
\title{Download tibble of the schedule (with results for completed games) for a given NBA team (use the 3 letter id)}
\usage{
get_team_schedule(team, season = getOption("tidynbadata.current_season"),
  version = getOption("tidynbadata.msf_version_id"))
}
\arguments{
\item{team}{the 3 letter id for an NBA team, use 'ALL-TEAMS' for a data.frame with the schedule for all teams}

\item{season}{the season you want the schedule for}

\item{version}{the msf api version}

\item{...}{additional paramaters passed to \code{msf_get_results()}}
}
\value{
a tibble
}
\description{
Download tibble of the schedule (with results for completed games) for a given NBA team (use the 3 letter id)
}
