% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{interpret_team}
\alias{interpret_team}
\title{Given a team input of unknown type (id, name, abbreviation) return all 3}
\usage{
interpret_team(team_input)
}
\description{
Given a team input of unknown type (id, name, abbreviation) return all 3
}
