% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pbp.R
\name{load_pbp}
\alias{load_pbp}
\title{Load an archived play-by-play file if it exists, otherwise,
process one and return it.}
\usage{
load_pbp(game_id, team = NA)
}
\arguments{
\item{game_id}{unique id for game}

\item{team}{any of team name, city, id or abbreviation. If NA returns a
list with a pbp for each team}
}
\value{
when team is NA a list with both team's pbp data is returned, when a
team_id is provided a data.frame of that team's pbp data is returned
}
\description{
Load an archived play-by-play file if it exists, otherwise,
process one and return it.
}
