% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_msf_pbp.R
\name{process_msf_pbp}
\alias{process_msf_pbp}
\title{High level function for converting raw msf pbp to a tidynbadata pbp tibble}
\usage{
process_msf_pbp(raw_msf_pbp)
}
\arguments{
\item{raw_msf_pbp}{a raw api pull of a play-by-play from mysportsfeeds
(as returned by \code{get_raw_pbp(game_id)})}
}
\value{
a length 2 list of pbp tibbles, one for each team involved.
This list is also archived.
}
\description{
Typically it is not necessary to call this function directly. It is better to
use \code{load_pbp(game_id, team)} which will handle acquisition of the
raw data and archiving of the results.
}
