% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{tidynbadata}
\alias{tidynbadata}
\title{tidynbadata constant environment}
\format{An object of class \code{environment} of length 6.}
\usage{
tidynbadata
}
\description{
A cache environment to store ours constants that are referenced in package functions.
}
\details{
use \code{ls(tidynbadata)} to see a list of currently available objects in this
environment
}
\keyword{datasets}
